requirements.txt
attrdict==2.0.1
matplotlib==3.7.5
numpy==1.24.3
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
Pillow==10.4.0
scikit_learn==1.3.2
scipy==1.10.1
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
pyyaml==6.0.2
pandas==2.0.3
casadi==3.6.7
tensorboard==2.13.0
gymnasium==1.1.0
